+++ src/komm/_error_control_block/SimplexCode.py
@@ -73,7 +73,7 @@ class SimplexCode(SystematicBlockCode):
     extended: bool = False
 
     @cached_property
-    def parity_submatrix(self) -> npt.NDArray[np.int_]:
+    def parity_submatrix(self) -> npt.NDArray[np.integer]:
         return hamming_parity_submatrix(self.kappa, self.extended).T
 
     @cache
